Add Bypass parser
Bypass (>=) allows you to compose a rule with an action to perform some
side effect but still return the rule's result to sebsequent parsers in
the pipeline.

Because of the extra power Bypass allows, Actions now return the value
of the action block, instead of just an empty string. Note that if that
value is nil, it could cause the parsing pipeline to fail.